Training/requirements.txt
torch>=2.0.0
torchvision
numpy
matplotlib
pillow